+++ weblate/accounts/models.py
@@ -783,6 +783,14 @@ class Profile(models.Model):
         verbose_name=gettext_lazy("Hide source if a secondary translation exists"),
         default=False,
     )
+    wide_tables = models.BooleanField(
+        verbose_name=gettext_lazy("Show all columns in lists using horizontal scrolling"),
+        default=False,
+        help_text=gettext_lazy(
+            "Instead of hiding columns on narrow screens, keep all columns and "
+            "scroll the table horizontally."
+        ),
+    )
     editor_link = models.CharField(
         default="",
         blank=True,